Temperature: Value from line is actually a int.

+++ data_logger/data_input/temperature.py
@@ -17,7 +17,7 @@ class Temperature(sensor.Sensor):
 
 	def _parse_temp_from_lines(self):
 		m = re.search("t=(?P<temp>-?\d+)", self.lines[1])
-		return float(m.group("temp")) / 1000.0
+		return int(m.group("temp")) / 1000.0
 
 	def get_data(self):
 		path = self.args["path"]
